# internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RSS Discovery API",
    "version": "1.0.0",
    "description": "Public feed catalog, search, discovery, and full-text cache API. Authenticate with Bearer tokens (rd_<level>_<secret>). Levels: readonly, standard, priority, admin."
  },
  "servers": [{"url": "/", "description": "This host"}],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "rd_<level>_<secret>"
      },
      "apiKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    },
...
  },
  "security": [{"bearerAuth": []}, {"apiKeyHeader": []}],
  "paths": {
...
    "/v1/discover/bulk": {
      "post": {
        "summary": "Bulk discover URLs (async job)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "urls": {"type": "array", "items": {"type": "string"}},
                  "refresh": {"type": "boolean"}
                }
              }
            }
          }
        },
        "responses": {"202": {"description": "Job accepted"}}
      }
    },
...
    "/v1/jobs/{id}": {
      "get": {
        "summary": "Get job status and items",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Job"}, "404": {"description": "Not found"}}
      }
    },
    "/v1/feeds/{id}/status": {
      "get": {
        "summary": "Per-feed freshness status vs SLO",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Status"}}
      }
    },
...
    "/v1/webhooks": {
      "get": {"summary": "List webhooks", "responses": {"200": {"description": "Webhooks"}}},
      "post": {
        "summary": "Create signed webhook",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["url"],
                "properties": {
                  "url": {"type": "string"},
                  "events": {"type": "string"},
                  "feed_id": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {"201": {"description": "Created with secret once"}}
      }
    },
    "/v1/webhooks/{id}": {
      "delete": {
        "summary": "Delete webhook",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Deleted"}}
      }
    },
...
    "/v1/websub/{id}/subscribe": {
      "post": {
        "summary": "Discover hub and subscribe for feed id",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"202": {"description": "Accepted"}}
      }
    },
    "/v1/websub/{id}": {
      "delete": {
        "summary": "Unsubscribe WebSub for feed",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Unsubscribed"}}
      }
    },
...
    "/v1/admin/oauth/clients/{id}": {
      "delete": {
        "summary": "Revoke OAuth client (admin)",
        "parameters": [{"name": "id", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Revoked"}}
      }
    }
  }
}